Move scaleImage() from UMSUtils to ImagesUtil
