+++ src/main/java/net/pms/util/UMSUtils.java
@@ -30,9 +30,6 @@ import java.text.Collator;
 import java.util.*;
 import java.util.List;
 import javax.imageio.ImageIO;
-import net.coobird.thumbnailator.Thumbnails;
-import net.coobird.thumbnailator.filters.Canvas;
-import net.coobird.thumbnailator.geometry.Positions;
 import net.pms.PMS;
 import net.pms.configuration.RendererConfiguration;
 import net.pms.dlna.*;
@@ -585,71 +582,5 @@ public class UMSUtils {
 			}
 			return null;
 		}
-	}
-
-	/**
-	 * @see #scaleImage(byte[], int, int, boolean)
-	 * @deprecated
-	 */
-	public static byte[] scaleImage(byte[] image, int width, int height, boolean outputBlank) {
-		return scaleImage(image, width, height, outputBlank, null);
-	}
-
-	/**
-	 * Creates a black background with the exact dimensions specified, then
-	 * centers the image on the background, preserving the aspect ratio.
-	 *
-	 * @param image
-	 * @param width
-	 * @param height
-	 * @param outputBlank whether to return null or a black image when the
-	 *                    image parameter is null
-	 * @param renderer
-	 *
-	 * @return the scaled image
-	 */
-	public static byte[] scaleImage(byte[] image, int width, int height, boolean outputBlank, RendererConfiguration renderer) {
-		ByteArrayInputStream in = null;
-		if (image == null && !outputBlank) {
-			return null;
-		} else if (image != null) {
-			in = new ByteArrayInputStream(image);
-		}
-
-		try {
-			BufferedImage img;
-			if (in != null) {
-				img = ImageIO.read(in);
-			} else {
-				img = new BufferedImage(width, height, BufferedImage.TYPE_INT_ARGB);
-			}
-
-			if (img == null) { // ImageIO doesn't support the image format
-				return null;
-			}
-
-			ByteArrayOutputStream out = new ByteArrayOutputStream();
-			if (renderer != null && renderer.isThumbnailPadding()) {
-				Thumbnails.of(img)
-					.size(width, height)
-					.addFilter(new Canvas(width, height, Positions.CENTER, Color.BLACK))
-					.outputFormat("JPEG")
-					.outputQuality(1.0f)
-					.toOutputStream(out);
-			} else {
-				Thumbnails.of(img)
-					.size(width, height)
-					.outputFormat("JPEG")
-					.outputQuality(1.0f)
-					.toOutputStream(out);
-			}
-
-			return out.toByteArray();
-		} catch (IOException e) {
-			LOGGER.debug("Failed to resize image: {}", e.getMessage());
-			LOGGER.trace("", e);
-		}
-
-		return null;
 	}
 }
